частичный рефракторинг + рефракторинг sql скриптов (flyway)

+++ src/main/resources/db/migration/V02__create_users_profiles_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS users_profiles (
     user_id    BIGINT PRIMARY KEY REFERENCES users(id) ON DELETE CASCADE,
     name       VARCHAR(255),
     birth_date DATE,
-    weight     DOUBLE PRECISION,
+    weight     NUMERIC(5,1),
     height     DOUBLE PRECISION,
     gender     VARCHAR(10) NOT NULL CHECK (gender IN ('MALE', 'FEMALE'))
 );
