+++ src/main/resources/db/migration/V12__create_sets_table.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS sets (
     id          BIGSERIAL PRIMARY KEY,
     session_id  BIGINT NOT NULL REFERENCES workout_sessions(id) ON DELETE CASCADE,
     exercise_id BIGINT NOT NULL REFERENCES exercises(id) ON DELETE CASCADE,
-    weight      DOUBLE PRECISION NOT NULL,
+    weight      NUMERIC(5,1) NOT NULL,
     reps        INTEGER NOT NULL,
     created_at  TIMESTAMP
 );
